The provided diff shows changes made to two shell scripts: `commit.sh` and `desc.sh`.
In `commit.sh`, two lines have been added at the beginning of the file:
```
#!/bin/bash

```
These lines indicate that the script should be executed using the Bash shell.

In `desc.sh`, two lines have also been added at the beginning of the file:
```
#!/bin/bash

```
Again, these lines indicate that the script should be executed using the Bash shell.

Additionally, in `desc.sh`, the script now retrieves the latest tag using the `git tag` command, sorts them in version order using `sort -V`, and selects the last tag using `tail -n 1`. The resulting tag is stored in the `latest_tag` variable.

The `commit_range` variable is then set to the range between the `latest_tag` and the `HEAD` commit.

Finally, the script can be used to generate a git commit message for the changes made.