The changes made to the commit.sh script involve modifying the command that generates the git commit message. Previously, the output of the `git diff --staged` command was piped directly into the sgpt command, which used the "gpt-4" model to generate a commit message based on the changes.
In the updated version of the script, the output of the sgpt command is further piped into the `sed` command before being written to the .commit file. The `sed` command is used to remove the leading and trailing double quotes from the commit message. This means that the commit message will no longer be enclosed in double quotes in the .commit file.