feat(setup.sh): add new setup script to automate environment setup refactor(config.fish): remove pip and vim aliases as they are now handled in setup.sh script
The setup.sh script is added to automate the process of setting up the development environment. This includes installing necessary packages, setting default python and pip versions, installing gnome extensions, and more. The pip and vim aliases are removed from the fish config file as they are now handled in the setup.sh script.