refactor(rc.lua): refactor tag creation to support different layouts for different screens
The code in rc.lua has been refactored to support different layouts for different screens. Previously, the tags were created with the default layout `awful.layout.layouts[1]`. Now, the code has been modified to check the outputs of each screen and set a different layout (`awful.layout.suit.tile.bottom`) for the screen with the output "DVI-I-2-2". This change improves the flexibility and customization of the Awesome window manager configuration.