chore(config.fish): update aliases and remove unused scripts feat(screen.sh): add script to manage screen configurations
The `config.fish` file was modified to update aliases and remove unused scripts. The `s` alias was changed to point to the `screen.sh` script instead of the `screen_auto_off.sh` script. The `rotate.sh`, `screen_auto_off.sh`, `screen_left.sh`, and `screen_normal.sh` scripts were deleted.

A new script `screen.sh` was added to manage screen configurations. It accepts command line arguments to set the screen to normal mode, rotate mode, duplicate mode, or toggle screen auto off. The script uses `xrandr` to configure the screens based on the provided arguments.

The purpose of these changes is to improve the usability and flexibility of the shell environment by updating aliases and removing unused scripts, as well as adding a new script to manage screen configurations.